Full pet and image cred on frontend and backend using AWS

+++ backend/src/db/migrations/02-shelters.ts
@@ -37,10 +37,10 @@ module.exports = {
         allowNull: false
       },
       lat: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(10, 8),
       },
       lon: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(10, 8),
       },
       phone: {
         type: Sequelize.STRING(14),
